requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
python-decouple>=3.8
Pillow>=10.0
scikit-learn>=1.3
pandas>=2.0
numpy>=1.24
joblib>=1.3
django-cors-headers>=4.0
django-filter>=23.0
gunicorn>=21.0
whitenoise>=6.0
